Rename htmlPlugin to fixLokiRefs
Rename Vite plugin file and reference to file in vite.config.js
