+++ fixLokiRefs.js
